# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Patent Intelligence API",
    "description": "Agentic RAG for multi-source patent intelligence. Research assistance only: not legal advice.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/patents/search": {
      "post": {
        "tags": [
          "patents"
        ],
        "summary": "Search Patents",
        "operationId": "search_patents_patents_search_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatentSearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatentSearchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/patents/import": {
      "post": {
        "tags": [
          "patents"
        ],
        "summary": "Import Patent",
        "operationId": "import_patent_patents_import_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatentImportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatentImportResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/patents/{patent_id}": {
      "get": {
        "tags": [
          "patents"
        ],
        "summary": "Get Patent",
        "operationId": "get_patent_patents__patent_id__get",
        "parameters": [
          {
            "name": "patent_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Patent Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatentOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "patents"
        ],
        "summary": "Delete Patent",
        "operationId": "delete_patent_patents__patent_id__delete",
        "parameters": [
          {
            "name": "patent_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Patent Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/analysis/invention": {
      "post": {
        "tags": [
          "invention analysis"
        ],
        "summary": "Analyze Invention",
        "description": "Takes ~1\u20133 minutes with local models: features, search, then one check per\nfeature \u00d7 document.",
        "operationId": "analyze_invention_analysis_invention_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InventionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventionAnalysisOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "invention analysis"
        ],
        "summary": "List Analyses",
        "operationId": "list_analyses_analysis_invention_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/InventionListItem"
                  },
                  "title": "Response List Analyses Analysis Invention Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/analysis/invention/{run_id}": {
      "get": {
        "tags": [
          "invention analysis"
        ],
        "summary": "Get Analysis",
        "operationId": "get_analysis_analysis_invention__run_id__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventionAnalysisOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/analysis/invention/{run_id}/report.md": {
      "get": {
        "tags": [
          "invention analysis"
        ],
        "summary": "Download Report",
        "operationId": "download_report_analysis_invention__run_id__report_md_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "InventionAnalysisOut": {
        "properties": {
          "run_id": {
            "type": "string",
            "format": "uuid",
            "title": "Run Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "latency_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latency Ms"
          },
          "result": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result"
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "status",
          "created_at",
          "latency_ms",
          "result"
        ],
        "title": "InventionAnalysisOut"
      },
      "InventionListItem": {
        "properties": {
          "run_id": {
            "type": "string",
            "format": "uuid",
            "title": "Run Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "features": {
            "type": "integer",
            "title": "Features"
          },
          "top_candidate": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Top Candidate"
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "status",
          "created_at",
          "description",
          "features",
          "top_candidate"
        ],
        "title": "InventionListItem"
      },
      "InventionRequest": {
        "properties": {
          "description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 8000
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "The invention in your own words"
          },
          "document_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Document Id",
            "description": "Or: an uploaded draft (its abstract and claim 1 are used)"
          },
          "use_patent_search": {
            "type": "boolean",
            "title": "Use Patent Search",
            "description": "Also search configured patent databases (EPO)",
            "default": true
          },
          "max_candidates": {
            "type": "integer",
            "maximum": 8.0,
            "minimum": 1.0,
            "title": "Max Candidates",
            "default": 5
          }
        },
        "type": "object",
        "title": "InventionRequest"
      },
...
      "PatentImportRequest": {
        "properties": {
          "source": {
            "type": "string",
            "maxLength": 32,
            "minLength": 1,
            "title": "Source"
          },
          "publication_number": {
            "type": "string",
            "maxLength": 64,
            "minLength": 3,
            "title": "Publication Number"
          }
        },
        "type": "object",
        "required": [
          "source",
          "publication_number"
        ],
        "title": "PatentImportRequest"
      },
      "PatentImportResponse": {
        "properties": {
          "patent": {
            "$ref": "#/components/schemas/PatentOut"
          },
          "already_imported": {
            "type": "boolean",
            "title": "Already Imported"
          }
        },
        "type": "object",
        "required": [
          "patent",
          "already_imported"
        ],
        "title": "PatentImportResponse"
      },
      "PatentOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "publication_number": {
            "type": "string",
            "title": "Publication Number"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "abstract": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Abstract"
          },
          "applicants": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Applicants"
          },
          "inventors": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Inventors"
          },
          "cpc_codes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Cpc Codes"
          },
          "publication_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Publication Date"
          },
          "filing_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Filing Date"
          },
          "priority_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Priority Date"
          },
          "family_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Family Id"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          },
          "has_claims": {
            "type": "boolean",
            "title": "Has Claims"
          },
          "has_description": {
            "type": "boolean",
            "title": "Has Description"
          },
          "chunk_count": {
            "type": "integer",
            "title": "Chunk Count"
          },
          "sections_found": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Sections Found"
          },
          "fetched_at": {
            "type": "string",
            "format": "date-time",
            "title": "Fetched At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "source",
          "publication_number",
          "title",
          "abstract",
          "applicants",
          "inventors",
          "cpc_codes",
          "publication_date",
          "filing_date",
          "priority_date",
          "family_id",
          "url",
          "has_claims",
          "has_description",
          "chunk_count",
          "sections_found",
          "fetched_at"
        ],
        "title": "PatentOut"
      },
      "PatentResultOut": {
        "properties": {
          "source": {
            "type": "string",
            "title": "Source"
          },
          "publication_number": {
            "type": "string",
            "title": "Publication Number"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "abstract": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Abstract"
          },
          "applicants": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Applicants"
          },
          "inventors": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Inventors"
          },
          "cpc_codes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Cpc Codes"
          },
          "publication_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Publication Date"
          },
          "filing_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Filing Date"
          },
          "family_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Family Id"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          },
          "imported_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Imported Id"
          },
          "similarity": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Similarity",
            "description": "Cosine similarity to the reference invention (not legal)"
          },
          "also_published_as": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Also Published As"
          }
        },
        "type": "object",
        "required": [
          "source",
          "publication_number",
          "title",
          "abstract",
          "applicants",
          "inventors",
          "cpc_codes",
          "publication_date",
          "filing_date",
          "family_id",
          "url",
          "imported_id"
        ],
        "title": "PatentResultOut"
      },
      "PatentSearchRequest": {
        "properties": {
          "keywords": {
            "type": "string",
            "maxLength": 500,
            "title": "Keywords",
            "default": ""
          },
          "cpc": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 10,
            "title": "Cpc"
          },
          "applicant": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Applicant"
          },
          "date_from": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date From"
          },
          "date_to": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date To"
          },
          "limit": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Limit",
            "default": 25
          },
          "sources": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sources",
            "description": "Default: all configured"
          },
          "dedup": {
            "type": "boolean",
            "title": "Dedup",
            "description": "Merge members of the same patent family",
            "default": true
          },
          "rank_against_document_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rank Against Document Id",
            "description": "Rank results by semantic similarity to this document"
          },
          "rank_against_patent_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rank Against Patent Id"
          }
        },
        "type": "object",
        "title": "PatentSearchRequest"
      },
      "PatentSearchResponse": {
        "properties": {
          "results": {
            "items": {
              "$ref": "#/components/schemas/PatentResultOut"
            },
            "type": "array",
            "title": "Results"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/SourceOutcomeOut"
            },
            "type": "array",
            "title": "Sources"
          },
          "deduplicated": {
            "type": "integer",
            "title": "Deduplicated",
            "default": 0
          },
          "reference_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reference Label"
          }
        },
        "type": "object",
        "required": [
          "results",
          "sources"
        ],
        "title": "PatentSearchResponse"
      },
...
      "SourceOutcomeOut": {
        "properties": {
          "source": {
            "type": "string",
            "title": "Source"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "returned": {
            "type": "integer",
            "title": "Returned"
          },
          "query_string": {
            "type": "string",
            "title": "Query String"
          },
          "cached": {
            "type": "boolean",
            "title": "Cached"
          },
          "latency_ms": {
            "type": "integer",
            "title": "Latency Ms"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "source",
          "total",
          "returned",
          "query_string",
          "cached",
          "latency_ms",
          "error"
        ],
        "title": "SourceOutcomeOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}